Remove BDB block database support
