+++ src/makefile.unix
@@ -4,7 +4,6 @@
 
 USE_UPNP:=0
 USE_IPV6:=1
-USE_LEVELDB:=1
 
 LINK:=$(CXX)
 
@@ -128,7 +127,9 @@ OBJS= \
     obj/util.o \
     obj/wallet.o \
     obj/walletdb.o \
-    obj/noui.o
+    obj/noui.o \
+    obj/leveldb.o \
+    obj/txdb.o
 
 
 all: bitcoind
@@ -139,17 +140,12 @@ test check: test_bitcoin FORCE
 #
 # LevelDB support
 #
-ifdef USE_LEVELDB
 LIBS += $(CURDIR)/leveldb/libleveldb.a $(CURDIR)/leveldb/libmemenv.a
-DEFS += $(addprefix -I,$(CURDIR)/leveldb/include) -DUSE_LEVELDB
+DEFS += $(addprefix -I,$(CURDIR)/leveldb/include)
 DEFS += $(addprefix -I,$(CURDIR)/leveldb/helpers)
-OBJS += obj/leveldb.o obj/txdb-leveldb.o
 leveldb/libleveldb.a:
 	@echo "Building LevelDB ..."; cd leveldb; make libleveldb.a libmemenv.a; cd ..;
 obj/leveldb.o: leveldb/libleveldb.a
-else
-OBJS += obj/txdb-bdb.o
-endif
 
 # auto-generated dependencies:
 -include obj/*.P
